+++ plugins/polio/js/src/domains/VaccineModule/SupplyChain/types.ts
@@ -7,6 +7,8 @@ import {
 } from '../../../../../../../hat/assets/js/apps/Iaso/types/utils';
 import { Vaccine } from '../../../constants/types';
 
+export type ScanResultStatus = 'CLEAN' | 'INFECTED' | 'ERROR' | 'PENDING';
+
 export type TabValue = 'vrf' | 'arrival_reports' | 'pre_alerts';
 
 export type VRF = {
@@ -49,7 +51,7 @@ export type PreAlert = {
     to_delete?: boolean;
     document?: File;
     can_edit: boolean;
-    scan_result?: string;
+    scan_result?: ScanResultStatus;
     scan_timestamp?: number;
 };
